refactor : 도메인 별 패키지 분리
