+++ src/main/java/subway/line/Line.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.line;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
